+++ R/v_outer.R
@@ -19,13 +19,19 @@
 #' #|------------------------------------------------------|
 #' #|    SETTING UP VARIOUS FUNCTIONS THAT WILL BE USED    |
 #' #|------------------------------------------------------|
-#' pooled.sd <- function(x, y) {
+#' pooled_sd <- function(x, y) {
 #'     n1 <- length(x)
 #'     n2 <- length(y)
 #'     s1 <- sd(x)
 #'     s2 <- sd(y)
 #'     sqrt(((n1-1)*s1 + (n2-1)*s2)/((n1-1) + (n2-1)))
 #' }
+#' 
+#' ## Effect Size: Cohen's d 
+#' cohens_d <- function(x, y) {
+#'     (mean(y) - mean(x))/pooled_sd(x, y)
+#' }
+#' 
 #' 
 #' ## Euclidean Distance
 #' euc_dist <- function(x,y) sqrt(sum((x - y) ^ 2))
@@ -39,7 +45,7 @@
 #' 
 #' ## A data.frame
 #' v_outer(mtcars, cor)
-#' v_outer(mtcars, pooled.sd)
+#' v_outer(mtcars, pooled_sd)
 #' v_outer(mtcars[, 1:7], euc_dist)
 #' v_outer(mtcars[, 1:7], sum2)
 #' v_outer(mtcars[, 1:7], arbitrary)
@@ -48,9 +54,11 @@
 #' mtcars2 <- lapply(mtcars[, 1:7], "[")
 #' v_outer(mtcars2, cor)
 #' v_outer(mtcars2, cor,  method = "spearman")
-#' v_outer(mtcars2, pooled.sd)
-#' print(v_outer(mtcars[, 1:7], pooled.sd), digits = 1)
-#' print(v_outer(mtcars[, 1:7], pooled.sd), digits = NULL)
+#' v_outer(mtcars2, pooled_sd)
+#' v_outer(split(mtcars[["mpg"]], mtcars[["carb"]]), cohens_d)
+#' v_outer(split(CO2[["uptake"]], CO2[["Plant"]]), cohens_d)
+#' print(v_outer(mtcars[, 1:7], pooled_sd), digits = 1)
+#' print(v_outer(mtcars[, 1:7], pooled_sd), digits = NULL)
 #' v_outer(mtcars2, euc_dist)
 #' v_outer(mtcars2, sum2)
 #' v_outer(mtcars2, arbitrary)